requirements.txt
bottle==0.12.13
Pillow==5.3.0
